Mirror Matchmaking Fully Functional
Completed Matchmaking using Mirror.

When a player tries to join a room that doesn't exist they will be told via panel that shows and hides after 3 seconds.

When a player joins then 3 more players can join they by using the same room name and port number. Both the room name and port number must match to join.

The connected players status's will update in the room scene displaying everyone's enterred username and if they are ready or unready.

When all players are ready (minimum of 2 and max of 4) then the game scene loads and the players are spawned.

If the host leavesthen all the players are disconnected abck to the matchmaking scene.

In the room scene the currently connected room name and port number will display so players can communicate with their friends on what to enter to join.

Scripts gracefully handle all player joins and leaves so that the room player list is constantly updated.

Tested and confirmed that two games can run simultaneously via different rooms